inference/requirements.txt
insightface
onnxruntime
gradio
numpy
Pillow
huggingface_hub
